requirements.txt
binance==0.3
click==8.1.3
defer==1.0.4
matplotlib==3.6.2
mplfinance==0.12.9b5
MySQL-python==1.2.5
mysql_connector_repackaged==0.3.1
nltk==3.7
numpy==1.23.5
pandas==1.5.2
pandas_datareader==0.10.0
requests==2.28.1
scikit_learn==1.1.3
scipy==1.9.3
SQLAlchemy==1.4.44
statsmodels==0.13.5
twint==2.1.20
yfinance==0.1.87
